[#44] Fix: 폴더명 수정
